requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.24.1
pydantic==2.10.0
pydantic-settings==2.7.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pyopenssl==24.0.0
cryptography==42.0.0
requests==2.31.0

validate-docbr==1.10.0

qrcode==7.4.2
APScheduler==3.10.4
supabase==2.3.0
python-dotenv==1.0.0
crcmod==1.7
reportlab==4.0.9
